# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BizzHive Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new buyer or seller account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
        "409":
          description: Email already exists
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
...
  /auth/delete-account:
    post:
      operationId: deleteAccount
      tags: [auth]
      summary: Delete or anonymize the current user's account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountBody"
      responses:
        "200":
          description: Account deleted or anonymized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteAccountResponse"
        "401":
          description: Not authenticated or incorrect password
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{id}:
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /orders/{id}/report-issue:
    post:
      operationId: reportOrderIssue
      tags: [orders]
      summary: Buyer raises a dispute instead of confirming delivery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportIssueBody"
      responses:
        "200":
          description: Dispute raised
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportIssueResponse"
        "400":
          description: Order is not awaiting confirmation
        "404":
          description: Order not found
...
components:
  schemas:
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          description: >-
            At least 8 characters with an uppercase letter, a lowercase letter,
            a number and a symbol. Must not be a common password or contain the
            user's own name or email.
        firstName:
          type: string
        lastName:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [buyer, seller]
          default: buyer
        phone:
          type: string
        vendorName:
          type: string
        vendorBio:
          type: string
        vendorLocation:
          type: string
      required: [email, password, firstName, lastName, role]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        displayName:
          type: ["string", "null"]
        role:
          type: string
          enum: [buyer, seller, admin]
        vendorId:
          type: ["integer", "null"]
        avatar:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
      required: [id, email, role]
    DeleteAccountBody:
      type: object
      properties:
        password:
          type: string
      required: [password]
    DeleteAccountResponse:
      type: object
      properties:
        status:
          type: string
          enum: [deleted, anonymized]
        message:
          type: string
      required: [status]
...
    CartItem:
      type: object
      properties:
        id:
          type: integer
        itemType:
          type: string
          enum: [course, product, session]
        itemId:
          type: integer
        title:
          type: string
        thumbnail:
          type: ["string", "null"]
        price:
          type: number
          format: float
        currency:
          type: string
        vendorId:
          type: integer
        vendorName:
          type: string
        # Present only for session line items.
        startsAt:
          type: ["string", "null"]
          format: date-time
        durationMinutes:
          type: ["integer", "null"]
      required: [id, itemType, itemId, title, price, currency, vendorName]
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
          format: float
        currency:
          type: string
        itemCount:
          type: integer
      required: [items, total, currency, itemCount]
    AddToCartBody:
      type: object
      properties:
        itemType:
          type: string
          enum: [course, product, session]
        itemId:
          type: integer
      required: [itemType, itemId]
...
    ReportIssueBody:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
    ReportIssueResponse:
      type: object
      properties:
        status:
          type: string
      required: [status]
...